Created a derivation system for different kinds of AoEs
