+++ Assets/Scripts/Misc_/DerivedDirectional.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 827aa38791c8f4a44bba2d90c540db0f
